[FIX] MRP 실행 예정일 계산 로직 수정 및 자재 주문 기준 단위 배수로 주문

+++ src/main/java/com/sampoom/factory/api/material/dto/MaterialResponseDto.java
@@ -17,7 +17,7 @@ public class MaterialResponseDto {
     private String materialCode;
     private Long materialCategoryId;
     private String materialCategoryName;
-    private Long quantity;
+    private Double quantity;
 
     public static MaterialResponseDto from(MaterialProjection material) {
         return MaterialResponseDto.builder()
@@ -30,7 +30,7 @@ public class MaterialResponseDto {
                 .build();
     }
 
-    public MaterialResponseDto withQuantity(Long quantity) {
+    public MaterialResponseDto withQuantity(Double quantity) {
         this.quantity = quantity;
         return this;
     }
